+++ examples/static_router/pages/about.rb
@@ -1,7 +1,13 @@
 class About < Porous::Page
   def self.route; ['about']; end
 
-  def html
-    '<h1>Porous Web Engine</h1><ul><li>About</li><li><a href="/">Home</a></li><li><a href="/contact">Contact</a></li>'
+  render :html do
+    h1 { 'Porous Web Engine' }
+
+    ul do
+      li { a(href: '/') { 'Home' } }
+      li { 'About' }
+      li { a(href: '/contact') { 'Contact' } }
+    end
   end
 end