+++ examples/static_router/pages/contact.rb
@@ -1,7 +1,13 @@
 class Contact < Porous::Page
   def self.route; ['contact']; end
 
-  def html
-    '<h1>Porous Web Engine</h1><ul><li>Contact</li><li><a href="/">Home</a></li><li><a href="/about">About</a></li>'
+  render :html do
+    h1 { 'Porous Web Engine' }
+
+    ul do
+      li { a(href: '/') { 'Home' } }
+      li { a(href: '/about') { 'About' } }
+      li { 'Contact' }
+    end
   end
 end