+++ lib/porous/cli.rb
@@ -8,10 +8,20 @@ module Porous
     def dev
       puts "Starting development server in #{`pwd`}..."
 
-      puts "Loading components…"
-      Dir['./components/**/*.rb'].each { |f| require f }
-      puts "Loading pages…"
-      Dir['./pages/**/*.rb'].each { |f| require f }
+      if Dir['./components/**/*.rb'].any?
+        puts "Loading components…"
+        Dir['./components/**/*.rb'].each { |f| require f }
+      else
+        puts "No components found!"
+      end
+
+      if Dir['./pages/**/*.rb'].any?
+        puts "Loading pages…"
+        Dir['./pages/**/*.rb'].each { |f| require f }
+      else
+        puts "No pages found!"
+      end
+
       puts "Ready!\n"
 
       Porous::Server.new.start