+++ lib/porous/page.rb
@@ -1,9 +1,6 @@
 module Porous
-  class Page
+  class Page < Component
     # Defaults to root route ('/') – override for each subclass
     def self.route; []; end
-
-    # By default it renders nothing
-    def html; nil; end
   end
 end