+++ spec/component_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe Porous::Component do
+  let(:component) { Porous::Component.new }
+end