+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Unreleased]
 
-## [0.1.0] - 2024-02-09
+- Simple Rackup web server
+- Extremely simple route matching to Pages
+- Only HTML rendering
+- No Entity support
+- No Event support
+- No persistence
+- No client-side (transpiled) components
+
+## [0.1.0]
 
 - Initial release